Fix showing error message if file does not exist

+++ setup/tmux.sh
@@ -1,3 +1,6 @@
 #!/bin/sh
-mv -f ~/.tmux.conf ~/.dotfiles/dotfiles_old/tmux.conf
+if [ -e ~/.tmux.conf ]
+then
+    mv -f ~/.tmux.conf ~/.dotfiles/dotfiles_old/tmux.conf
+fi
 ln -s ~/.dotfiles/tmux.conf ~/.tmux.conf
