+++ setup/vim.sh
@@ -1,3 +1,6 @@
 #!/bin/sh
-mv -f ~/.vimrc ~/.dotfiles/dotfiles_old/vimrc
+if [ -e ~/.vimrc ]
+then
+    mv -f ~/.vimrc ~/.dotfiles/dotfiles_old/vimrc
+fi
 ln -s ~/.dotfiles/vimrc ~/.vimrc
