+++ setup/zsh.sh
@@ -4,11 +4,20 @@ then
     sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 fi
 
-mv -f ~/.zshrc ~/.dotfiles/dotfiles_old/zshrc
+if [ -e ~/.zshrc ]
+then
+    mv -f ~/.zshrc ~/.dotfiles/dotfiles_old/zshrc
+fi
 ln -s ~/.dotfiles/zshrc ~/.zshrc
 mkdir -p ~/.oh-my-zsh/custom/themes
 mkdir -p ~/.dotfiles/dotfiles_old/oh-my-zsh/custom/themes
-mv -f ~/.oh-my-zsh/custom/themes ~/.dotfiles/dotfiles_old/oh-my-zsh/custom/themes
+if [ -e ~/.oh-my-zsh/custom/themes ]
+then
+    mv -f ~/.oh-my-zsh/custom/themes ~/.dotfiles/dotfiles_old/oh-my-zsh/custom/themes
+fi
 ln -s ~/.dotfiles/oh-my-zsh/custom/themes/ ~/.oh-my-zsh/custom
-mv -f ~/.oh-my-zsh/custom/alias.zsh ~/.dotfiles/dotfiles_old/oh-my-zsh/custom/alias.zsh
+if [ -e ~/.oh-my-zsh/custom/alias.zsh ]
+then
+    mv -f ~/.oh-my-zsh/custom/alias.zsh ~/.dotfiles/dotfiles_old/oh-my-zsh/custom/alias.zsh
+fi
 ln -s ~/.dotfiles/oh-my-zsh/custom/alias.zsh ~/.oh-my-zsh/custom/alias.zsh
